feat: combine aug smoketests into single file to allow for cleanup
Otherwise, you cannot cleanup the base template.
